src/backend/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
Django==4.1.3
django-environ==0.9.0
djangorestframework==3.14.0
dnspython==2.2.1
et-xmlfile==1.1.0
idna==3.4
numpy==1.23.5
openpyxl==3.0.10
pandas==1.5.2
pymongo==4.3.3
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
